Maintenance: Enhances release workflow
Adds write permissions for content and id-token to the publish job.

This allows the workflow to create and push a release branch, as well as update the package version and commit the changes.

Also configures the checkout action to fetch the entire git history in order for npm version to correctly determine the next version.